fix: move handlers to metric domain
