+++ src/main/java/com/aws/greengrass/clientdevices/auth/metrics/handlers/SessionCreationEventHandler.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package com.aws.greengrass.clientdevices.auth.session.handlers;
+package com.aws.greengrass.clientdevices.auth.metrics.handlers;
 
 import com.aws.greengrass.clientdevices.auth.api.DomainEvents;
 import com.aws.greengrass.clientdevices.auth.metrics.ClientDeviceAuthMetrics;
